Правки главной страницы и страницы просмотра лота
